refactor: move x icon to the right folders and update imports
